refactor: XYNE-282 comments resolved

+++ server/api/admin.ts
@@ -1835,7 +1835,7 @@ export const GetAdminChats = async (c: Context) => {
         break
       case "cost":
         orderByClause = desc(
-          sql`COALESCE(SUM(CASE WHEN ${messages.deletedAt} IS NULL THEN ${messages.cost} ELSE 0 END), 0)::numeric`,
+          sql`COALESCE(SUM(CASE WHEN ${messages.deletedAt} IS NULL THEN ${messages.cost} ELSE 0 END), 0)`,
         )
         break
       case "tokens":
